make space in path compatible
Inline::Filters passed an unquoted filepath to the CC of the location of
the .c file, if there is a space in the path, the CC can't find the file
and errors out. "C:\Documents and Settings\Owner\.cpan\build" is a common
path on Win32 where things are built from with a space in the path, so
I::F was uninstallable using cpanm or cpan shell before this commit.